Move ZoneRepositoryInterface from the Bundle to the Component
