+++ src/Sylius/Component/Addressing/Repository/ZoneRepositoryInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\AddressingBundle\Repository;
+namespace Sylius\Component\Addressing\Repository;
 
 use Sylius\Component\Addressing\Model\AddressInterface;
 use Sylius\Component\Addressing\Model\ZoneInterface;
