+++ src/Sylius/Component/Addressing/spec/Matcher/ZoneMatcherSpec.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace spec\Sylius\Component\Addressing\Matcher;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\AddressingBundle\Repository\ZoneRepositoryInterface;
 use Sylius\Component\Addressing\Model\AddressInterface;
 use Sylius\Component\Addressing\Model\ZoneInterface;
+use Sylius\Component\Addressing\Repository\ZoneRepositoryInterface;
 
 final class ZoneMatcherSpec extends ObjectBehavior
 {
